Add a hapi sample
- Modelled as a hapi plugin. There's very little point trying
to wrap the functionality to a hapi plugin which could be used
as-is, so this just a sample on how to integrate to one's hapi
app as a self-crafted plugin.
- hapi requires currently node >=4.0.0 and uses available ES6
features extensively, thus this sample is written with the same
approach - in other words, won't work on node v0.x.y.
- Certificates, images and styles are the same as in the express
sample.