Remove sample app dependency from unit tests
The point of the unit test for 'tupas/ok' route is to verify
that it will send an error event if verifying query parameters
fails (with empty POST payload). There's no need to use sample
app for this (it could well have its own set of tests), better
to verify this piece of core functionality with a self-contained
test setup.