Don't bind to a server app automatically
If wanting to use this module client-side for form rendering,
having the dependency to server-side route handling forced
handing in a dummy implementation of appHandler in create().
This can be avoided with this simple change which makes the
appHandler optional. The .bindHandlers() can be called also
separately if need be.